Hide 'post_date' input field on 'Add/Edit Recipe' form
Set input field for 'post_date' to the Bootstrap class: 'd-none' to hide input field from user. Since the date input is autofilled using JS, the user is not required to fill in this field and renders the display of the field, redundant.

The field is set to readonly required with the style attribute: 'pointer-events: none;' for now in case Bootstrap fails to load the 'd-none' class or if it is decided to show the input field in the future.