Edit Recipe function
Created function to allow for editing recipes and updating db.

Removed JS function to insert date into post_date field on add/edit recipe form due to an error when iterating recipe data by id.

This error prevented the date being inserted into the field for future iterations, only the first iteration did it work.

To solve this, datetime was imported into routes.py and a function was created to get today's date.

This date is assigned to the post_date variable in the 'add_recipe' and 'edit_recipe' functions.

This ensures that the date is always updated in all iterations of the recipes table and in all instances where it is displayed in HTML via Jinja templates.