Create Back-End for 'Rate Recipe' Form
Updated 'models.py' with new columns for 'rating' and 'average_rating'.

'rating' is structured as an array of integers while 'average_rating' is structured as a float.

Added a route in 'routes.py' for the 'rate_recipe' form.

This route contains a function which updates the 'rating' column with the new value appended to the array, modifying the array each time the form is submitted.

When the form is submitted, another function calculates the mean average of all values in the 'rating' array and assigns the sum to the 'average_rating' column.

The 'average_rating' is displayed in 'recipe.html' using Jinja2 conditional statements to display star icons based on the rating out of 5.

The 'average_rating' is displayed next to these icons. The float is rounded to 1 decimal place.