User Session Functionality
Created model for User in models.py

Created a Sign In and Sign Up page for users to create accounts and sign in to access features such as adding cuisine categories and recipes.

Defensive programming includes only displaying edit/delete buttons if the user currently logged into the session, matches the owner or user_id assigned to that recipe/category.

The rate recipe form is hidden from users if the user currently logged in matches the owner of the recipe. This prevents users from rating their own recipes.